Core: PageHeading and layout.master versatility, plus a FindControlRecursive method extension for Control
Created PageHeading class which inherits from PageTitle. This is simply an alias.
Removed styleshout reference from generic layout.master
Added (protected) isNonCmsBasePage property to layout.master.cs. If page is not "isNonCmsBasePage", the currentPage property is populated using the CacheHelper.GetPage method with the pageid in the query string. This will help skin developers use the currentPage object in the layout.master on pages like the blog's viewpost.aspx