Core: Avatar control updates
Added LinkTo... options in hopes to make this control a bit friendlier to skin developers while maintaing backwards compatibility. The overall goal is to be able to link to a user's public or private profile from the skin, even if gravatar is used and have those links when mojo is installed in a virtual directory or the site is a folder-based tenant.