Core: Removed query that isn't supported by SQLite.
You can't rename columns of a table in SQLite, you have to create a temp table, delete the old table, recreate the table with the column name change you want, and then copy the data out of the temp table.

I have a feeling I did this manually initially and forgot to clean up the repo.

And can I just say that SQLite is stupid for not supporting basic SQL queries. lol